requirements.txt
asgiref==3.7.2
Django==3.2.25
django-js-asset==2.0.0
django-mptt==0.14.0
Pillow
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.7.1
